requirements.txt
PyYAML>=6.0
Jinja2>=3.1.2
requests>=2.28.0
argparse>=1.4.0
Groq
python-dotenv